Split bonespirit overlay in two
Previously bonespirit overlay contained some of my personal
configuration files. These files were moved to a separate private
overlay. This is a public part of the original bonespirit overlay.

Public overlay's name is the same as the original one. However, all
users of the original overlay need to re-add this overlay manually
(layman -d bonespirit && layman -a bonespirit).

My apologies for the inconvenience.